Split up src/schema/mongoose_schema_utils.ts in to sub modules
